Chỉnh sửa thêm service đổ dữ liệu sử dụng cho toàn hệ thống
