+++ ProjectNews/src/app/component/catalog-detail/catalog-detail.component.ts
@@ -3,6 +3,7 @@ import {HttpClient} from "@angular/common/http";
 import * as xml2js from "xml2js";
 import {RSSNews} from "../../model/NewsRSS";
 import {ActivatedRoute, Router} from "@angular/router";
+import {DataService} from "../../service/data.service";
 
 @Component({
   selector: 'app-catalog-detail',
@@ -12,28 +13,11 @@ import {ActivatedRoute, Router} from "@angular/router";
 export class CatalogDetailComponent implements OnInit, OnDestroy {
   RssData: RSSNews;
 
-  constructor(private http: HttpClient, private route: ActivatedRoute, private router: Router) {
+  constructor(private http: HttpClient, private route: ActivatedRoute, private router: Router, private feed: DataService) {
     route.params.subscribe(val => {
       let id = this.route.snapshot.params.id;
-      this.GetRssFeedData(id)
+      this.RssData = feed.GetRssFeedData(id)
     })
-  }
-
-  GetRssFeedData(paramater: string) {
-    const requestOptions: Object = {
-      observe: "body",
-      responseType: "text"
-    };
-    this.http
-      .get<any>("https://thethao247.vn/" + paramater + ".rss", requestOptions)
-      .subscribe(data => {
-        let parseString = xml2js.parseString;
-        parseString(data, (err, result: RSSNews) => {
-          this.RssData = result;
-        });
-
-      });
-    return this.RssData
   }
 
   forRange(start: number, end: number) {
